oCheat somewhat working now
managed to have oCheat somewhat working with the N and P keys going and going to the previous room. the only problem is when pressing P at room01 it'll crash the game as there won't be anywhere to go to afterwards. also changed chest color to blue and found flickering issue to be the chest sprite having an animation. still need to look into aspect issue and why the player isn't spawning.